backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.19
google-genai==0.8.0
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.24.0
google-cloud-storage==2.19.0
